+++ leetcode/src/lc0426_partition_equal_subset_sum.cpp
@@ -25,19 +25,19 @@ bool LC0426_PartitionEqualSubsetSum_DP1::canPartition(std::vector<int>& nums)
         return false;
     }
 
-    int32_t target = sum / 2;
-    if (target < *maxItr) {
+    uint32_t target = (uint32_t)sum / 2;
+    if (target < (uint32_t)*maxItr) {
         return false;
     }
 
     // i 数组当前的长度
     // j 刚好达到的值
-    vector<vector<bool>> dp(numCnt, vector<bool>(target + 1, false));
+    vector<vector<bool>> dp(numCnt, vector<bool>((uint32_t)(target + 1), false));
 
     dp[0][(uint32_t)nums[0]] = true;
     for (uint32_t i = 1; i < numCnt; i++) {
-        int32_t n = nums[i];
-        for (int32_t j = 0; j <= target; j++) {
+        uint32_t n = (uint32_t)nums[i];
+        for (uint32_t j = 0; j <= target; j++) {
             if (j == 0) {
                 dp[i][j] = true;
                 continue;
@@ -56,7 +56,7 @@ bool LC0426_PartitionEqualSubsetSum_DP1::canPartition(std::vector<int>& nums)
 
 bool LC0426_PartitionEqualSubsetSum_DP2::canPartition(std::vector<int>& nums)
 {
-    int32_t numCnt = (uint32_t)nums.size();
+    uint32_t numCnt = (uint32_t)nums.size();
     if (numCnt < 2) {
         return false;
     }
@@ -67,17 +67,17 @@ bool LC0426_PartitionEqualSubsetSum_DP2::canPartition(std::vector<int>& nums)
     if (sum % 2 == 1) {
         return false;
     }
-    int32_t target = sum / 2;
-    if (target < *maxItr) {
+    uint32_t target = (uint32_t)sum / 2;
+    if (target < (uint32_t)*maxItr) {
         return false;
     }
     vector<bool> dp(target + 1, false);
 
     dp[0] = true;
-    for (int32_t i = 1; i < numCnt; i++) {
-        int32_t n = nums[i];
+    for (uint32_t i = 1; i < numCnt; i++) {
+        uint32_t n = (uint32_t)nums[i];
         dp[n] = true;
-        for (int32_t j = target; j > n; --j) {
+        for (uint32_t j = target; j > n; --j) {
             dp[j] = dp[j] || dp[j - n];
         }
     }
